+++ Karakuri Library/Karakuri/KRFont.mm
@@ -11,18 +11,18 @@
     @method KRFont
     Constructor
  */
-KRFont::KRFont(const std::string& fontName, float size)
+KRFont::KRFont(const std::string& fontName, double size)
 {
     mFontName = fontName;
 
     NSString *fontNameStr = [[NSString alloc] initWithCString:fontName.c_str() encoding:NSUTF8StringEncoding];
 
 #if KR_MACOSX || KR_IPHONE_MACOSX_EMU
-    mFontObj = [[NSFont fontWithName:fontNameStr size:size] retain];
+    mFontObj = [[NSFont fontWithName:fontNameStr size:(float)size] retain];
 #endif
     
 #if KR_IPHONE && !KR_IPHONE_MACOSX_EMU
-    mFontObj = [[UIFont fontWithName:fontNameStr size:size] retain];
+    mFontObj = [[UIFont fontWithName:fontNameStr size:(float)size] retain];
 #endif
     
     [fontNameStr release];
